+++ src/callbacksCustom/postPlayerFatalDamage.ts
@@ -1,5 +1,0 @@
-import { racePostPlayerFatalDamage } from "../features/race/callbacks/postPlayerFatalDamage";
-
-export function main(player: EntityPlayer): boolean | void {
-  return racePostPlayerFatalDamage(player);
-}