+++ src/features/optional/major/fastTravel/callbacks/postGameStartedContinued.ts
@@ -1,4 +1,4 @@
-import { enableAllInputs } from "../../../../mandatory/disableAllInputs";
+import { enableInputs } from "../../../../../../../isaacscript-common/dist";
 import { FastTravelState } from "../enums";
 import v from "../v";
 
@@ -6,6 +6,6 @@ export default function fastTravelPostGameStartedContinued(): void {
   // Cancel fast-travel if we save & quit in the middle of the jumping animation
   if (v.run.state === FastTravelState.FADING_TO_BLACK) {
     v.run.state = FastTravelState.DISABLED;
-    enableAllInputs();
+    enableInputs();
   }
 }