+++ src/features/race/callbacks/preCustomRevive.ts
@@ -1,0 +1,5 @@
+import * as seededDeath from "../seededDeath";
+
+export default function racePreCustomRevive(player: EntityPlayer): int | void {
+  return seededDeath.preCustomRevive(player);
+}