+++ src/features/race/types/RevivalType.ts
@@ -1,0 +1,4 @@
+enum RevivalType {
+  SEEDED_DEATH,
+}
+export default RevivalType;